projects/daily-task-assistant/requirements.txt
anthropic>=0.30
python-dotenv>=1.0
pyyaml>=6.0
pytest>=7.0
fastapi>=0.110
uvicorn[standard]>=0.23
firebase-admin>=7.1.0
google-auth>=2.23

